scraper/base_scraper.py
import asyncio
import aiohttp
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import logging
from typing import Dict, List, Optional, Union
import json
import os
from datetime import datetime

class BaseScraper:
    def __init__(self, base_url: str, use_selenium: bool = True):
        self.base_url = base_url
        self.use_selenium = use_selenium
        self.session = None
        self.driver = None
        self.setup_logging()
        
    def setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
        self.logger = logging.getLogger(__name__)

    async def initialize(self):
        if self.use_selenium:
            chrome_options = Options()
            chrome_options.add_argument('--headless')
            chrome_options.add_argument('--no-sandbox')
            chrome_options.add_argument('--disable-dev-shm-usage')
            self.driver = webdriver.Chrome(options=chrome_options)
        else:
            self.session = aiohttp.ClientSession()

    async def close(self):
        if self.driver:
            self.driver.quit()
        if self.session:
            await self.session.close()

    async def get_page_content(self, url: str) -> str:
        if self.use_selenium:
            self.driver.get(url)
            WebDriverWait(self.driver, 10).until(
                EC.presence_of_element_located((By.TAG_NAME, "body"))
            )
            return self.driver.page_source
        else:
            async with self.session.get(url) as response:
                return await response.text()

    def parse_profile(self, html_content: str) -> Dict:
        soup = BeautifulSoup(html_content, 'html.parser')
        profile_data = {
            'name': self._extract_name(soup),
            'phone': self._extract_phone(soup),
            'location': self._extract_location(soup),
            'age': self._extract_age(soup),
            'date': self._extract_date(soup),
            'message': self._extract_message(soup),
            'images': self._extract_images(soup),
            'metadata': {
                'scraped_at': datetime.now().isoformat(),
                'source_url': self.base_url
            }
        }
        return profile_data

    def _extract_name(self, soup: BeautifulSoup) -> Optional[str]:
        # Implement name extraction logic
        pass

    def _extract_phone(self, soup: BeautifulSoup) -> Optional[str]:
        # Implement phone extraction logic
        pass

    def _extract_location(self, soup: BeautifulSoup) -> Optional[str]:
        # Implement location extraction logic
        pass

    def _extract_age(self, soup: BeautifulSoup) -> Optional[int]:
        # Implement age extraction logic
        pass

    def _extract_date(self, soup: BeautifulSoup) -> Optional[str]:
        # Implement date extraction logic
        pass

    def _extract_message(self, soup: BeautifulSoup) -> Optional[str]:
        # Implement message extraction logic
        pass

    def _extract_images(self, soup: BeautifulSoup) -> List[str]:
        # Implement image extraction logic
        pass

    async def scrape_profiles(self, states: List[str] = None) -> List[Dict]:
        if states is None:
            states = ['all']
        
        profiles = []
        try:
            await self.initialize()
            # Implement state-based scraping logic
            for state in states:
                state_profiles = await self._scrape_state(state)
                profiles.extend(state_profiles)
        finally:
            await self.close()
        
        return profiles

    async def _scrape_state(self, state: str) -> List[Dict]:
        # Implement state-specific scraping logic
        pass

    def save_profiles(self, profiles: List[Dict], output_dir: str = 'data'):
        os.makedirs(output_dir, exist_ok=True)
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        output_file = os.path.join(output_dir, f'profiles_{timestamp}.json')
        
        with open(output_file, 'w', encoding='utf-8') as f:
            json.dump(profiles, f, ensure_ascii=False, indent=2)
        
        self.logger.info(f"Saved {len(profiles)} profiles to {output_file}") 