requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
pandas>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
face-recognition>=1.3.0
networkx>=3.1
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
tqdm>=4.66.0
scikit-learn>=1.3.0
cmake>=3.25.0
dlib>=19.24.0 